// apps/main/src/models/v2/dataplane/byStatusQueryParamParameterItem.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */

export type ByStatusQueryParamParameterItem =
  (typeof ByStatusQueryParamParameterItem)[keyof typeof ByStatusQueryParamParameterItem];

export const ByStatusQueryParamParameterItem = {
  executing: 'executing',
  failed: 'failed',
  succeeded: 'succeeded',
} as const;