// apps/main/src/models/v2/dataplane/deploymentPage.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { DeploymentSummary } from './deploymentSummary';

/**
 * A page of deployments
 */
export interface DeploymentPage {
  items: DeploymentSummary[];
  /** The page token to use to request the next page of items */
  next_page_token?: string;
}