// apps/main/src/models/v2/dataplane/deploymentSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { DeploymentId } from './deploymentId';
import type { DeploymentMetrics } from './deploymentMetrics';
import type { OrganizationId } from './organizationId';

/**
 * Summary information for a deployment, more information can be retrieved through the GetDeployment APIs.
 */
export interface DeploymentSummary {
  org_id: OrganizationId;
  /** The Project id which the environment exists in */
  project_id: string;
  /** The Environment id in the project which is being deployed */
  env_id: string;
  id: DeploymentId;
  /** The time the deployment was created */
  created_at: string;
  /** The user that created the deployment */
  created_by: string;
  /** The time the deployment succeeded, failed, or terminated */
  completed_at?: string;
  /** The mode of the deployment. */
  mode: string;
  /** Whether the deployment is executed in plan mode. */
  plan_only: boolean;
  /** When mode is 'rollback', this field will hold the uuid of the deployment rollback to. */
  rollback_to_deployment_id?: string;
  /** The status of the deployment. */
  status: string;
  /** A message associated with the 'status'. */
  status_message: string;
  metrics: DeploymentMetrics;
}